refactor(web): move register components to a dedicated folder
