+++ packages/web/components/Register/index.js
@@ -1,0 +1,6 @@
+export { default as StepOne } from './StepOne';
+export { default as StepTwo } from './StepTwo';
+export { default as StepThree } from './StepThree';
+export { default as StepFour } from './StepFour';
+export { default as StepFive } from './StepFive';
+export { default as steps } from './steps';
