+++ packages/web/components/Register/steps.js
