+++ packages/web/screens/Register/styles.js
@@ -145,168 +145,3 @@ export const PageTitle = styled.h3`
 		}
 	`}
 `;
-
-export const StepTitle = styled.h3`
-	${({ theme: { colors } }) => css`
-		color: ${colors.black};
-		font-weight: 500;
-		margin-bottom: 0.8rem;
-	`}
-`;
-
-export const StepSubtitle = styled.p`
-	${({ theme: { colors } }) => css`
-		color: ${colors.lightGray2};
-		margin-bottom: 4rem;
-	`}
-`;
-
-export const RegisterTypeTitle = styled.h3`
-	${({ theme: { colors } }) => css`
-		color: ${colors.lightGray2};
-		font-size: 2.8rem;
-		font-weight: 500;
-		line-height: 3.3rem;
-		margin-bottom: 4rem;
-
-		> p {
-			font-size: 1.6rem;
-			line-height: 3.3rem;
-		}
-	`}
-`;
-
-export const Actions = styled.div`
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-`;
-
-export const InputsWrapper = styled.div`
-	> div:not(:last-child) {
-		margin-bottom: 2.4rem;
-	}
-`;
-
-export const CheckboxLabel = styled.p`
-	${({ theme: { colors } }) => css`
-		color: ${colors.black};
-		font-weight: 500;
-		font-size: 1.2rem;
-	`}
-`;
-
-export const CheckboxWrapper = styled.div`
-	margin-bottom: 4rem;
-	> div:first-child {
-		margin-bottom: 0.8rem;
-	}
-`;
-
-export const FloatingAction = styled.div`
-	${({ theme: { screens, metrics }, justifyContent }) => css`
-		display: flex;
-		justify-content: space-between;
-		flex-direction: row-reverse;
-		margin: auto -3.2rem -3.2rem -3.2rem;
-
-		> button,
-		a {
-			border-radius: 00;
-		}
-
-		> button + #back-button {
-			display: none;
-		}
-
-		@media screen and (min-width: ${screens.medium - 1}px) {
-			margin: 0;
-			margin-top: 17.6rem;
-			display: flex;
-			justify-content: ${justifyContent || 'space-between'};
-
-			> button,
-			a {
-				padding: 0.2rem 0.8rem;
-				border-radius: ${metrics.baseRadius}rem;
-			}
-
-			> button + #back-button {
-				display: inherit;
-			}
-		}
-	`}
-`;
-
-export const Form = styled.form`
-	${({ theme: { screens } }) => css`
-		display: flex;
-		flex-direction: column;
-		height: 100%;
-
-		@media screen and (min-width: ${screens.medium + 1}px) {
-			max-width: 41.5rem;
-		}
-	`}
-`;
-
-export const StepInfo = styled.p`
-	${({ theme: { colors } }) => css`
-		color: ${colors.silver};
-		margin-bottom: 2.4rem;
-	`}
-`;
-
-export const VerificationCodeWrapper = styled.div`
-	${({ theme: { screens } }) => css`
-		margin: 5.6rem auto 0;
-		width: fit-content;
-		> div {
-			gap: 0.8rem;
-			justify-content: center;
-		}
-
-		@media screen and (min-width: ${screens.medium + 1}px) {
-			margin: 5.6rem 0 0;
-			> div {
-				gap: 1.6rem;
-			}
-		}
-	`}
-`;
-
-export const ResendEmailLink = styled.div`
-	margin-top: 0.4rem;
-
-	button {
-		width: fit-content;
-		margin-left: auto;
-		font-size: 1.2rem;
-		font-weight: 500;
-	}
-`;
-
-export const FinishContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: space-evenly;
-	height: 100%;
-	text-align: center;
-`;
-
-export const FinishTitle = styled.h3`
-	${({ theme: { colors } }) => css`
-		color: ${colors.secondary};
-		max-width: 27.5rem;
-		line-height: 3rem;
-	`}
-`;
-
-export const FinishSubtitle = styled.p`
-	${({ theme: { colors } }) => css`
-		color: ${colors.silver};
-		font-size: 1.6rem;
-		line-height: 2.4rem;
-	`}
-`;
